setup: Install ninja
Ninja can't be installed as a jhbuild module, for the unsatisfying reason
that its download consists of a zip file with a bare binary in it, not in
a subdirectory. JHbuild doesn't know how to deal with archives without
subdirectories!

Instead, we download and install ninja as part of setup.