Build GNOME 3.24 platform
We switch from 3.22 to the current stable release of GNOME, 3.24. We can
remove js24, and we need to add pulseaudio to macos-sysdeps since it was
turned into a system dependency in the upstream module set.

Configuration overrides are needed so that Mozilla's custom stuff can find
a Python interpreter with ssl support, and for e-d-s.